# [xy] Speed up bigquery destination in data integration pipeline.