… mage_integrations/mage_integrations/destinations/sql/utils.py
… mage_integrations/mage_integrations/tests/destinations/test_base.py