# Description
<!-- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context.
List any dependencies that are required for this change.
-->
Speed up bigquery destination in data integration pipeline.
* Only run schema creation command, table creation and alter commands for the first subbatch in destination
* Allow maximum batch size to be configured via `maximum_batch_size_mb` key in destination config
* Remove redundant records loop through in bigquery destination when `use_batch_load` is true
* Speed up exporting datetime column to BigQuery by converting records to dataframe and use `pd.to_datetime` to convert datetime column
  * User PARQUET format for most of the cases
  * Use CSV format when there're json columns: https://stackoverflow.com/questions/76983677/json-format-not-supported-when-using-load-table-from-dataframe-bigquery
  * Use load_table_from_json when there's ARRAY type since REPEATED mode is not supported in CSV format

# How Has This Been Tested?
<!-- Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce.
-->
Config
```yaml
use_batch_load: true
maximum_batch_size_mb: 500
```

- [x] Tested exporting 50k rows to BigQuery. Time before optimization: 90s. Time after optimization: 20s.
* Load from dataframe (default to parquet): 20s
* Load from dataframe (csv format with schema): 20s
* Load from dataframe (csv format without schema): 41s
* Load from json: 42s


Column type tested
- [x] string
- [x] int
- [x] datetime
- [x] array
- [x] object


# Checklist
- [x] The PR is tagged with proper labels (bug, enhancement, feature, documentation)
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas

cc:
<!-- Optionally mention someone to let them know about this pull request -->